Simplify classes
