--- subscriptions/subscriptions-impl/src/main/java/com/duckduckgo/subscriptions/impl/ui/onboarding/SubscriptionOnboardingDetailsViewModel.kt
@@ -1,25 +1,0 @@
-/*
- * Copyright (c) 2026 DuckDuckGo
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.duckduckgo.subscriptions.impl.ui.onboarding
-
-import androidx.lifecycle.ViewModel
-import com.duckduckgo.anvil.annotations.ContributesViewModel
-import com.duckduckgo.di.scopes.FragmentScope
-import javax.inject.Inject
-
-@ContributesViewModel(FragmentScope::class)
-class SubscriptionOnboardingDetailsViewModel @Inject constructor() : ViewModel()
