--- subscriptions/subscriptions-impl/src/main/java/com/duckduckgo/subscriptions/impl/ui/onboarding/SubscriptionOnboardingOrigin.kt
@@ -1,24 +1,0 @@
-/*
- * Copyright (c) 2026 DuckDuckGo
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.duckduckgo.subscriptions.impl.ui.onboarding
-
-/**
- * Where the onboarding flow was entered from.
- * - [PURCHASE]: just bought a subscription — start at the intro (with confetti).
- * - [SETTINGS]: re-entered from Subscription Settings — skip the intro and resume at the last step.
- */
-enum class SubscriptionOnboardingOrigin { PURCHASE, SETTINGS }
