--- subscriptions/subscriptions-impl/src/main/java/com/duckduckgo/subscriptions/impl/ui/onboarding/SubscriptionOnboardingStepHost.kt
@@ -1,32 +1,0 @@
-/*
- * Copyright (c) 2026 DuckDuckGo
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.duckduckgo.subscriptions.impl.ui.onboarding
-
-/**
- * Implemented by the activity hosting the onboarding steps. A step fragment casts its host activity
- * to this and calls these to drive the flow.
- */
-interface SubscriptionOnboardingStepHost {
-    /** Mark the current step completed (persisted) and advance. */
-    fun onStepCompleted()
-
-    /** Advance without completing the current step. */
-    fun onNextStep()
-
-    /** Go back to the previous screen in this session. */
-    fun onBackStep()
-}
